move cookies func to cookies.js
